requirements.txt
torch>=2.0
torchaudio>=2.0
openai-whisper
pyannote.audio>=3.0.1
huggingface_hub
soundfile
